Moving all TypedDataset methods that cannot be optimized by Catalyst into a separate implicit class
