+++ dataset/src/test/scala/frameless/ops/unoptimized/FilterTests.scala
@@ -1,9 +1,11 @@
 package frameless
+package ops
+package unoptimized
 
 import org.scalacheck.Prop
 import org.scalacheck.Prop._
 
-class ForwardedFilterTests extends TypedDatasetSuite {
+class FilterTests extends TypedDatasetSuite {
   test("filter") {
     def prop[A: TypedEncoder](filterFunction: A => Boolean, data: Vector[A]): Prop =
       TypedDataset.create(data).filter(filterFunction).collect().run().toVector =? data.filter(filterFunction)
